Data.Vector.NonEmpty.Internal

Description

Internal module exposing the constructors for NonEmptyVector and NonEmptyMVector.

Warning: Since the constructors are exposed here, by using this module, you take on the risks that you break the non-emptiness invariants of the main modules. Use at your own risk.

Since: 0.2.1.0

Immutable boxed vectors

newtype NonEmptyVector a

NonEmptyVector is a thin wrapper around Vector that witnesses an API requiring non-empty construction, initialization, and generation of non-empty vectors by design.

A newtype wrapper was chosen so that no new pointer indirection is introduced when working with Vectors, and all performance characteristics inherited from the Vector API still apply.

Since: 0.2.1.0

Constructors

NonEmptyVector

Fields

_neVec :: Vector a

Mutable boxed vectors

newtype NonEmptyMVector s a

NonEmptyMVector is a thin wrapper around MVector that witnesses an API requiring non-empty construction, initialization, and generation of non-empty vectors by design.

A newtype wrapper was chosen so that no new pointer indirection is introduced when working with MVectors, and all performance characteristics inherited from the MVector API still apply.

Since: 0.2.1.0

Constructors

NonEmptyMVector

Fields

_nemVec :: MVector s a

Mutable vector aliases

type NonEmptyIOVector = NonEmptyMVector RealWorld

NonEmptyMVector parametrized by PrimState

Since: 0.2.1.0

type NonEmptySTVector s = NonEmptyMVector s

NonEmptyMVector parametrized by ST

Since: 0.2.1.0